Modified slider
The slider can now be set up with a power-law response…this means the
rate of change will depend on where you are on the slider. This is
useful for the polygons because it spreads out the lower side numbers.

An alternative constructor has been provided for this - the default
behaviour is linear.